Add files via upload
Write a program to check if two given String is Anagram of each other return true if Input
string is anagram otherwise return false. (By the way, two String is called anagram, if
they contain the same characters but in different order e.g. army and mary, stop and
pots, etc. Anagrams are actually a mix-up of characters in String.)